Separating and rebuilding `Results`.
